refactor: mbti 테스트 score 생성 로직 모듈화
